Refactored the era-reanalysis sub-pipeline into a module
