+++ src/dart_pipeline/constants.py
@@ -5,11 +5,10 @@ COMPRESSED_FILE_EXTS = [".tif.gz", ".tar.gz", ".tar.bz2", ".zip", ".7z"]
 
 INTEGER_PARAMS = ["year"]
 
-DEFAULT_SOURCES_ROOT = "data/sources"
-DEFAULT_OUTPUT_ROOT = "data/processed"
-DEFAULT_PLOTS_ROOT = "data/plots"
-
 BASE_DIR = Path(__file__).parent.parent.parent
+DEFAULT_SOURCES_ROOT = BASE_DIR / 'data' / 'sources'
+DEFAULT_OUTPUT_ROOT = BASE_DIR / 'data' / 'processed'
+DEFAULT_PLOTS_ROOT = BASE_DIR / 'data' / 'plots'
 
 COL_BLUE = "\033[0;34m"
 COL_CYAN = "\033[0;36m"
@@ -17,6 +16,9 @@ COL_OFF = "\033[0m"
 
 MSG_SOURCE = COL_BLUE + "  source" + COL_OFF
 MSG_PROCESS = COL_CYAN + " process" + COL_OFF
+
+# Smallest single-precision floating-point number
+MIN_FLOAT = -3.4028234663852886e38
 
 TERRACLIMATE_METRICS = [
     "aet",  # water_evaporation_amount_mm
